requirements.txt
numpy>=1.24
PyYAML>=6.0
matplotlib>=3.8
torch>=2.2
pytest>=8.0
jupyter>=1.0
